requirements.txt
certifi==2020.6.20
pyenchant==3.2.2
